+++ plugins/cost-insights/src/components/CostOverviewCard/CostOverviewChart.tsx
@@ -18,8 +18,6 @@ import { DateTime } from 'luxon';
 import { useTheme, Box } from '@material-ui/core';
 import {
   ComposedChart,
-  ContentRenderer,
-  TooltipProps,
   XAxis,
   YAxis,
   Tooltip as RechartsTooltip,
@@ -50,6 +48,8 @@ import { useCostOverviewStyles as useStyles } from '../../utils/styles';
 import { groupByDate, toDataMax, trendFrom } from '../../utils/charts';
 import { aggregationSort } from '../../utils/sort';
 import { CostOverviewLegend } from './CostOverviewLegend';
+import { TooltipRenderer } from '../../types';
+import { useConfig } from '../../hooks';
 
 type CostOverviewChartProps = {
   metric: Maybe<Metric>;
@@ -66,6 +66,7 @@ export const CostOverviewChart = ({
 }: CostOverviewChartProps) => {
   const theme = useTheme<CostInsightsTheme>();
   const styles = useStyles(theme);
+  const { baseCurrency } = useConfig();
 
   const data = {
     dailyCost: {
@@ -98,10 +99,7 @@ export const CostOverviewChart = ({
         : {}),
     }));
 
-  const tooltipRenderer: ContentRenderer<TooltipProps> = ({
-    label,
-    payload = [],
-  }) => {
+  const tooltipRenderer: TooltipRenderer = ({ label, payload = [] }) => {
     if (isInvalid({ label, payload })) return null;
 
     const dataKeys = [data.dailyCost.dataKey, data.metric.dataKey];
@@ -110,17 +108,18 @@ export const CostOverviewChart = ({
         ? DateTime.fromMillis(label)
         : DateTime.fromISO(label!);
     const title = date.toUTC().toFormat(DEFAULT_DATE_FORMAT);
+    const formatGraphValueWith = formatGraphValue(baseCurrency);
     const items = payload
       .filter(p => dataKeys.includes(p.dataKey as string))
-      .map(p => ({
+      .map((p, i) => ({
         label:
           p.dataKey === data.dailyCost.dataKey
             ? data.dailyCost.name
             : data.metric.name,
         value:
           p.dataKey === data.dailyCost.dataKey
-            ? formatGraphValue(p.value as number, data.dailyCost.format)
-            : formatGraphValue(p.value as number, data.metric.format),
+            ? formatGraphValueWith(Number(p.value), i, data.dailyCost.format)
+            : formatGraphValueWith(Number(p.value), i, data.metric.format),
         fill:
           p.dataKey === data.dailyCost.dataKey
             ? theme.palette.blue
@@ -161,7 +160,7 @@ export const CostOverviewChart = ({
           <YAxis
             domain={[() => 0, 'dataMax']}
             tick={{ fill: styles.axis.fill }}
-            tickFormatter={formatGraphValue}
+            tickFormatter={formatGraphValue(baseCurrency)}
             width={styles.yAxis.width}
             yAxisId={data.dailyCost.dataKey}
           />
@@ -186,7 +185,6 @@ export const CostOverviewChart = ({
             dataKey="trend"
             dot={false}
             isAnimationActive={false}
-            label={false}
             strokeWidth={2}
             stroke={theme.palette.blue}
             yAxisId={data.dailyCost.dataKey}
@@ -196,7 +194,6 @@ export const CostOverviewChart = ({
               dataKey={data.metric.dataKey}
               dot={false}
               isAnimationActive={false}
-              label={false}
               strokeWidth={2}
               stroke={theme.palette.magenta}
               yAxisId={data.metric.dataKey}