+++ plugins/scaffolder/src/index.ts
@@ -43,6 +43,7 @@ export {
   ScaffolderFieldExtensions,
 } from './extensions';
 export type {
+  CustomFieldExtensionSchema,
   CustomFieldValidator,
   FieldExtensionOptions,
   FieldExtensionComponentProps,
@@ -61,6 +62,12 @@ export {
   scaffolderPlugin,
 } from './plugin';
 export * from './components';
+export {
+  rootRouteRef,
+  nextRouteRef,
+  selectedTemplateRouteRef,
+  nextSelectedTemplateRouteRef,
+} from './routes';
 export type { TaskPageProps } from './components/TaskPage';
 
 /** next exports */