+++ plugins/techdocs/src/reader/components/TechDocsReaderPageContent/dom.tsx
@@ -21,7 +21,7 @@ import { useTheme, useMediaQuery } from '@material-ui/core';
 
 import { BackstageTheme } from '@backstage/theme';
 import { CompoundEntityRef } from '@backstage/catalog-model';
-import { useApi } from '@backstage/core-plugin-api';
+import { useAnalytics, useApi } from '@backstage/core-plugin-api';
 import { scmIntegrationsApiRef } from '@backstage/integration-react';
 
 import {
@@ -63,6 +63,7 @@ export const useTechDocsReaderDom = (
   const isMobileMedia = useMediaQuery(MOBILE_MEDIA_QUERY);
   const sanitizerTransformer = useSanitizerTransformer();
   const stylesTransformer = useStylesTransformer();
+  const analytics = useAnalytics();
 
   const techdocsStorageApi = useApi(techdocsStorageApiRef);
   const scmIntegrationsApi = useApi(scmIntegrationsApiRef);
@@ -175,13 +176,20 @@ export const useTechDocsReaderDom = (
             // detect if CTRL or META keys are pressed so that links can be opened in a new tab with `window.open`
             const modifierActive = event.ctrlKey || event.metaKey;
             const parsedUrl = new URL(url);
+            const fullPath = `${parsedUrl.pathname}${parsedUrl.search}${parsedUrl.hash}`;
+
+            // capture link clicks within documentation
+            const linkText =
+              (event.target as HTMLAnchorElement | undefined)?.innerText || url;
+            const to = url.replace(window.location.origin, '');
+            analytics.captureEvent('click', linkText, { attributes: { to } });
 
             // hash exists when anchor is clicked on secondary sidebar
             if (parsedUrl.hash) {
               if (modifierActive) {
-                window.open(`${parsedUrl.pathname}${parsedUrl.hash}`, '_blank');
+                window.open(fullPath, '_blank');
               } else {
-                navigate(`${parsedUrl.pathname}${parsedUrl.hash}`);
+                navigate(fullPath);
                 // Scroll to hash if it's on the current page
                 transformedElement
                   ?.querySelector(`[id="${parsedUrl.hash.slice(1)}"]`)
@@ -189,9 +197,9 @@ export const useTechDocsReaderDom = (
               }
             } else {
               if (modifierActive) {
-                window.open(parsedUrl.pathname, '_blank');
+                window.open(fullPath, '_blank');
               } else {
-                navigate(parsedUrl.pathname);
+                navigate(fullPath);
               }
             }
           },
@@ -218,7 +226,7 @@ export const useTechDocsReaderDom = (
           onLoaded: () => {},
         }),
       ]),
-    [theme, navigate],
+    [theme, navigate, analytics],
   );
 
   useEffect(() => {