+++ packages/catalog-client/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @backstage/catalog-client
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@1.1.3
+  - @backstage/errors@1.1.3
+
+## 1.1.2-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@1.1.3-next.0
+  - @backstage/errors@1.1.3-next.0
+
 ## 1.1.1
 
 ### Patch Changes