+++ plugins/bitbucket-cloud-common/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @backstage/plugin-bitbucket-cloud-common
 
+## 0.2.1
+
+### Patch Changes
+
+- d089fbe7dc: Add interfaces for Bitbucket Cloud (webhook) events.
+- Updated dependencies
+  - @backstage/integration@1.4.0
+
+## 0.2.1-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/integration@1.4.0-next.0
+
 ## 0.2.0
 
 ### Minor Changes